requirements.txt
# requirements.txt content:
flask
flask-cors
web3
requests
python-dotenv
bcrypt
json
